refactor(relationships): extract reference vocabulary and scope-path homes [roadmap:rebuild-scale]
Implements a Movement-A item of rac/roadmaps/rebuild-scale.md.

Pure code motion: the section-vocabulary/reference-extraction
foundation moves to services/references.py and the filesystem-scope
helpers to services/scope_paths.py with public names, retiring both
private-symbol leaks (scope.py's three underscore imports and
rename.py's _build_resolution_index). All previous names keep resolving
via re-exports (ADR-062); behavior byte-frozen.
